model: Factor out the ScanSummary filtering
This allows to only filter the ScanSummary which will be used later.
